pyproject.toml
[tool.poetry]
name = "virtual-contributor-engine-expert"
version = "0.8.1"
description = "Alkemio Generative AI Virtul Persona"
authors = ["Alkemio BV <info@alkem.io>"]
license = "EUPL-1.2"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.11"
langchain = "0.2.7"
langchain-core= "0.2.37"
langchain-community = "0.2.7"
openai = "1.43.0"
load-dotenv = "0.1.0"
async-timeout = "4.0.3"
aio-pika = "9.3.1"
aiormq = "6.7.7"
chromadb-client = "0.4.25.dev0"
azure-ai-inference = "1.0.0b4"
langchain-openai = "0.1.23"
asyncio = "3.4.3"

[tool.poetry.group.dev.dependencies]
flake8 = "6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

